# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Conditions API expose CRUD actions to condition objects on servers",
        "title": "Condition orchestrator API",
        "contact": {}
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/servers/{uuid}/condition": {
            "get": {
                "description": "Returns the active/pending Condition for the serverID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "ConditionGet",
                "operationId": "conditionGet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server ID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Condition Kind",
                        "name": "conditionKind",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_metal-automata_conditionorc_pkg_api_v1_orchestrator_types.ServerResponse"
                        }
                    }
                }
            }
        },
        "/servers/{uuid}/condition-status/{conditionKind}/{conditionID}": {
            "put": {
                "description": "Publishes an update to the Condition StatusValue KV",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "ConditionStatusUpdate",
                "operationId": "conditionStatusUpdate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server ID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Condition Kind",
                        "name": "conditionKind",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Condition ID",
                        "name": "conditionID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Only update timestamp in the StatusValue entry",
                        "name": "ts_update",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_metal-automata_conditionorc_pkg_api_v1_orchestrator_types.ServerResponse"
                        }
                    }
                }
            }
        },
        "/servers/{uuid}/condition-task/{conditionKind}": {
            "get": {
                "description": "Queries a *rivets.Task object from KV for a condition\nControllers will not have always know the taskID and so this enables querying\nthe active Task object using the serverID, conditionKind parameters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "taskQuery returns the active/pending condition for a serverID",
                "operationId": "taskQuery",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server ID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Condition Kind",
                        "name": "conditionKind",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_metal-automata_conditionorc_pkg_api_v1_orchestrator_types.ServerResponse"
                        }
                    }
                }
            }
        },
        "/servers/{uuid}/condition-task/{conditionKind}/{conditionID}": {
            "post": {
                "description": "Publishes a *rivets.Task object to the KV for a condition",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "taskPublish",
                "operationId": "taskPublish",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server ID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Condition Kind",
                        "name": "conditionKind",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Condition ID",
                        "name": "conditionID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_metal-automata_conditionorc_pkg_api_v1_orchestrator_types.ServerResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "common.Status": {
            "type": "object",
            "properties": {
                "health": {
                    "type": "string"
                },
                "post_code": {
                    "type": "integer"
                },
                "post_code_status": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
...
        "condition.Condition": {
            "type": "object",
            "properties": {
                "client": {
                    "description": "Client is the user/jwt user that requested the condition.",
                    "type": "string"
                },
                "createdAt": {
                    "description": "CreatedAt is when this object was created.",
                    "type": "string"
                },
                "failOnCheckpointError": {
                    "description": "Should the worker executing this condition fail if its unable to checkpoint\nthe status of work on this condition.",
                    "type": "boolean"
                },
                "fault": {
                    "description": "Fault is used to introduce faults into the controller when executing on a condition.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/condition.Fault"
                        }
                    ]
                },
                "id": {
                    "description": "ID is the identifier for this condition.",
                    "type": "string"
                },
                "kind": {
                    "description": "Kind is one of Kind.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/condition.Kind"
                        }
                    ]
                },
                "parameters": {
                    "description": "Parameters is a JSON object that is agreed upon by the controller\nreconciling the condition and the client requesting the condition.",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "spanID": {
                    "description": "SpanID enables tracking a Condition and any associated Conditions",
                    "type": "string"
                },
                "state": {
                    "description": "State is one of State",
                    "allOf": [
                        {
                            "$ref": "#/definitions/condition.State"
                        }
                    ]
                },
                "status": {
                    "description": "Status is a JSON object that is agreed upon by the controller\nreconciling the condition and the client requesting the condition.",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "target": {
                    "description": "Target is the identifier for the target server this Condition is applicable for.",
                    "type": "string"
                },
                "traceID": {
                    "description": "TraceID enables tracking a Condition and any associated Conditions",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "UpdatedAt is when this object was last updated.",
                    "type": "string"
                },
                "version": {
                    "description": "Version identifies the revision number for this struct.",
                    "type": "string"
                }
            }
        },
        "condition.Fault": {
            "type": "object",
            "properties": {
                "delayDuration": {
                    "description": "Introduce specified delay in execution of the condition on the controller.\n\naccepts the string format of time.Duration - 5s, 5m, 5h",
                    "type": "string"
                },
                "failAt": {
                    "description": "FailAt is a controller specific task/stage that the condition should fail in execution.\n\nfor example, in the flasher controller, setting this field to `init` will cause the\ncondition task to fail at initialization.",
                    "type": "string"
                },
                "panic": {
                    "description": "will cause the condition execution to panic on the controller.",
                    "type": "boolean"
                }
            }
        },
...
        "condition.Kind": {
            "type": "string",
            "enum": [
                "inventory",
                "virtualMediaMount",
                "serverControl",
                "firmwareInstall",
                "firmwareInstallInband",
                "biosControl",
                "broker",
                "broker.acquireServer",
                "broker.releaseServer"
            ],
            "x-enum-varnames": [
                "Inventory",
                "VirtualMediaMount",
                "ServerControl",
                "FirmwareInstall",
                "FirmwareInstallInband",
                "BiosControl",
                "Broker",
                "BrokerAcquireServer",
                "BrokerReleaseServer"
            ]
        },
        "condition.State": {
            "type": "string",
            "enum": [
                "pending",
                "active",
                "failed",
                "succeeded"
            ],
            "x-enum-varnames": [
                "Pending",
                "Active",
                "Failed",
                "Succeeded"
            ]
        },
        "condition.StatusMsg": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                },
                "ts": {
                    "type": "string"
                }
            }
        },
        "condition.StatusRecord": {
            "type": "object",
            "properties": {
                "records": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/condition.StatusMsg"
                    }
                }
            }
        },
        "condition.Task-any-any": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "data": {
                    "description": "Data holds Condition Task specific data"
                },
                "facility_code": {
                    "description": "FacilityCode identifies the facility this task is to be executed in.",
                    "type": "string"
                },
                "fault": {
                    "description": "Fault is a field to inject failures into a flasher task execution,\nthis is set from the Condition only when the worker is run with fault-injection enabled.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/condition.Fault"
                        }
                    ]
                },
                "id": {
                    "description": "Task unique identifier, this is set to the Condition identifier.",
                    "type": "string"
                },
                "kind": {
                    "description": "Kind is the type of Condition this Task is derived from",
                    "allOf": [
                        {
                            "$ref": "#/definitions/condition.Kind"
                        }
                    ]
                },
                "parameters": {
                    "description": "Parameters holds Condition specific parameters for this task"
                },
                "server": {
                    "description": "Server holds attributes about target server this task is for.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.Server"
                        }
                    ]
                },
                "spanID": {
                    "type": "string"
                },
                "state": {
                    "description": "state is the state of the install",
                    "allOf": [
                        {
                            "$ref": "#/definitions/condition.State"
                        }
                    ]
                },
                "status": {
                    "description": "status holds informational data on the state",
                    "allOf": [
                        {
                            "$ref": "#/definitions/condition.StatusRecord"
                        }
                    ]
                },
                "task_version": {
                    "description": "StructVersion indicates the Task object version and is used to determine Task  compatibility.",
                    "type": "string"
                },
                "traceID": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "worker_id": {
                    "description": "WorkerID is the identifier for the worker executing this task.",
                    "type": "string"
                }
            }
        },
...
        "github_com_metal-automata_conditionorc_pkg_api_v1_orchestrator_types.ServerResponse": {
            "type": "object",
            "properties": {
                "condition": {
                    "$ref": "#/definitions/condition.Condition"
                },
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                },
                "task": {
                    "$ref": "#/definitions/condition.Task-any-any"
                }
            }
        },
...
        "types.Component": {
            "type": "object",
            "properties": {
                "attributes": {
                    "$ref": "#/definitions/types.ComponentAttributes"
                },
                "firmware": {
                    "$ref": "#/definitions/common.Firmware"
                },
                "id": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "serial": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/common.Status"
                },
                "updated": {
                    "type": "string"
                },
                "vendor": {
                    "type": "string"
                }
            }
        },
        "types.ComponentAttributes": {
            "type": "object",
            "properties": {
                "architecture": {
                    "type": "string"
                },
                "block_size_bytes": {
                    "type": "integer"
                },
                "bus_info": {
                    "type": "string"
                },
                "capabilities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/common.Capability"
                    }
                },
                "capable_speed_gbps": {
                    "type": "integer"
                },
                "capacity_bytes": {
                    "type": "integer"
                },
                "chassis_type": {
                    "type": "string"
                },
                "clock_speed_hz": {
                    "type": "integer"
                },
                "cores": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "drive_type": {
                    "type": "string"
                },
                "form_factor": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "interface_type": {
                    "type": "string"
                },
                "macaddress": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "negotiated_speed_gbps": {
                    "type": "integer"
                },
                "oem": {
                    "type": "boolean"
                },
                "oem_id": {
                    "type": "string"
                },
                "part_number": {
                    "type": "string"
                },
                "physid": {
                    "type": "string"
                },
                "power_capacity_watts": {
                    "type": "integer"
                },
                "product_name": {
                    "type": "string"
                },
                "protocol": {
                    "type": "string"
                },
                "size_bytes": {
                    "type": "integer"
                },
                "slot": {
                    "type": "string"
                },
                "smart_errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "smart_status": {
                    "type": "string"
                },
                "speed_bits": {
                    "type": "integer"
                },
                "speed_gbps": {
                    "type": "integer"
                },
                "storage_controller": {
                    "type": "string"
                },
                "supported_controller_protocol": {
                    "type": "string"
                },
                "supported_device_protocol": {
                    "type": "string"
                },
                "supported_raid_types": {
                    "type": "string"
                },
                "threads": {
                    "type": "integer"
                },
                "wwn": {
                    "type": "string"
                }
            }
        },
...
        "types.Server": {
            "type": "object",
            "properties": {
                "bios_cfg": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "bmc_address": {
                    "type": "string"
                },
                "bmc_password": {
                    "type": "string"
                },
                "bmc_user": {
                    "type": "string"
                },
                "components": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Component"
                    }
                },
                "facility": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "serial": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated": {
                    "type": "string"
                },
                "vendor": {
                    "type": "string"
                }
            }
        },
...
    }
}